@@ -9,22 +9,22 @@
         "@fortawesome/free-brands-svg-icons": "^6.1.2",
         "@fortawesome/free-solid-svg-icons": "^6.1.2",
         "@fortawesome/react-fontawesome": "^0.2.0",
-        "@types/react": "^18.0.17",
-        "@types/react-dom": "^18.0.6",
-        "@types/react-responsive": "^8.0.4",
-        "@types/react-router-dom": "^5.3.2",
+        "@types/react": "^18.2.12",
+        "@types/react-dom": "^18.2.5",
+        "@types/react-responsive": "^8.0.5",
+        "@types/react-router-dom": "^5.3.3",
         "@types/react-transition-group": "^4.4.4",
         "@types/uuid": "^8.3.1",
-        "axios": "^0.27.2",
+        "axios": "^1.4.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
-        "react-responsive": "^9.0.0-beta.4",
-        "react-router-dom": "^6.0.2",
-        "react-scripts": "^5.0.0",
+        "react-responsive": "^9.0.2",
+        "react-router-dom": "^6.12.1",
+        "react-scripts": "^5.0.1",
         "react-transition-group": "^4.4.2",
         "socket.io-client": "^4.3.2",
         "ts-node": "^10.4.0",
-        "typescript": "^4.4.4",
+        "typescript": "^5.1.3",
         "uuid": "^8.3.2"
     },
     "devDependencies": {
