@@ -8,16 +8,16 @@
   "dependencies": {
     "bcryptjs": "^2.4.3",
     "config": "^3.3.6",
-    "dotenv": "^16.0.0",
+    "dotenv": "^16.1.4",
     "express": "^4.17.1",
     "express-validator": "^6.13.0",
     "jsonwebtoken": "^9.0.0",
-    "mongoose": "^6.1.6",
+    "mongoose": "^7.2.4",
     "nodemailer": "^6.7.1",
     "path": "^0.12.7",
     "socket.io": "^4.3.2",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.4.4"
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.3"
   },
   "devDependencies": {
     "@types/bcryptjs": "^2.4.2",
